Reverted back to pre-meteor updates.
I didn’t like some of the changes I made to the project. Specificly, I tried to assign templates to all the objects on in the game and I think that was a big mistake. There is no way that is going to be preformant. I think a better approach is to have an animation loop running inside one of the templates and only live-query and reactive data for data coming from the server, like other platers commands.