Changed the view back to NOT rotate the screen
I’ve gone back and forth on this a couple of times. The original idea was that the screen was supposed to be like a ship’s console. So when the shop rotated the entire view would rotate so the ship’s heading would stay fixed. But as the game as eveloved into a more action oriented game it just become more intuitive, I think, to fix the screen and have the ship rotate. Note, however, that the ship is fixed to the cener of the screen. So you can’t fly off the screen in the way you could in say, Astroids.