Fixed Safari issue with canvas element layouts
It turned out that Safari does not like the canvas element to have inline style attributes. Worked fine in Chrome and Firefox… Who know…..

The fix was to put the attribute into an external style sheet.  Works, but, annoying because know it is not as obvious in the code what the intention is. Oh well.