FPS, local mode and updateID changes
Made changes to the server, client and egnine interfaces to all the client and server to independently set the engines current FPS. It used to be hardcoded to 60fps.

I added a local mode to the client to make it eiaser to toggle between local and multiplayer for testing.

I added a new updateID attribute to the updates that allows me to detect if updates are coming down to the client properly log results to the console.