Rendering changes to support larger map
Larger maps would not work because the larger they got the more stars I had to draw for the back ground. To fix this, I added code to only render stars if theyfall within the visual range of the current viewport.

I also updated the visual range to be basedo on the larger viewing side insted of the smaller viewing side.