Render Optimization
Cheanged the code to only look up a reference to the player's ship once per frame. It should even be that often but it was doing the look up like 10 times per frame.

Also change the code to only lookup the camera sides once per frame. It was doing it once per game object AND once per star. Per frame! That is a lot of unnecisary calculations!